clk: Inform the core about consumer devices
We'd like to have a pointer to the device that's consuming a particular
clk deep in the clk framework so we can link the consumer to the clk
provider with the PM device links. Add a device argument to
clk_hw_create_clk() for this so it can be used in subsequence patches to
add and remove the device link.