soc: qcom: cmd-db: Map with read-only mappings
The command DB is read-only already to the kernel because everything is
const marked once we map it. Let's go one step further and try to map
the memory as read-only in the page tables. This should make it harder
for random code to corrupt the database and change the contents.

Signed-off-by: Stephen Boyd <swboyd@chromium.org>