debugobjects: Allow debug_obj_descr to be const
The debugobject core could be slightly harder to corrupt if we make the
debug_obj_descr a pointer to const memory. Depending on the
architecture, const datastructures are placed into read-only memory and
thus are harder to corrupt or hijack. This descriptor is used to fix up
stuff like timers and workqueues when core kernel datastructures are
busted, so moving the descriptors to read-only memory will make
debugobjects more resilient to something going wrong and then corrupting
the function pointers inside struct debug_obj_descr.

Signed-off-by: Stephen Boyd <swboyd@chromium.org>