drm/msm/dp: Simplify 'rc' variable usage
This variable, the return variable, is usually initialized to 0, but
then overwritten with the return value of some other function. Instead
of initializing it to 0 all the time, let's reduce that and avoid using
gotos to exit functions when there aren't locks held or anything that
needs to be done on the error path out of functions. This makes code
slightly shorter and clearer by not having to track lifetime of a
variable or gotos while reading code.

Signed-off-by: Stephen Boyd <swboyd@chromium.org>